+++ lib/rack/lobster.rb
@@ -58,7 +58,7 @@ end
 
 if $0 == __FILE__
   require 'rack'
-  require 'rack/showexceptions'
+  require 'rack/show_exceptions'
   Rack::Server.start(
     :app => Rack::ShowExceptions.new(Rack::Lint.new(Rack::Lobster.new)), :Port => 9292
   )
