+++ test/spec_builder.rb
@@ -1,7 +1,7 @@
 require 'rack/builder'
 require 'rack/lint'
 require 'rack/mock'
-require 'rack/showexceptions'
+require 'rack/show_exceptions'
 require 'rack/urlmap'
 
 class NothingMiddleware
